[ext-code-space] Remove CodeDataContainer::code_cage_base field
... add pass the code cage base or Isolate explicitly to FromCodeT()
functions. In most of the cases we have either the code cage base or an
Isolate from which we can load the code cage base value. For the other
rare cases we deal with compiled code and thus we can compute the Code
object pointer from the code entry point by subtracting the Code header
size.

Drive-by: cleanup some unnecessary Code <-> CodeT conversions.

Bug: v8:11880
Change-Id: Ief12d13d28266c95eb8d5ba93b3afd568c4683cc