[heap-broker] Use MapRef for transition groups
Element access feedback transition groups were stored as a vector of
Handle<Map>, and converted to MapRef on access. We can just immediately
create MapRefs for this vector and skip the later conversions.

This also cleans up a couple of locations that thought they needed to
deal with null Maps -- these are already filtered out when creating the
transition group array.

Bug: v8:7700
Change-Id: I766b006f1523827d60f7626151199fc0eb99f881