[runtime] Simplify const tracking
To avoid requiring optimized code to do too extensive checking, this CL
reduces the extent to which we recognise similar values as the same
value for const fields. For smi/tagged/heap-object fields we only
support the exact same object; also if we might have a new reboxed
version of a double. For double fields we only support the exact same
bit patterns.

Change-Id: Ifd136aa442054fe9059d0de47e31455d6e1e25ca