Allow embedder to use script merging for streamed scripts
This CL adds new functions on StreamedSource that closely match the
functions already available on ConsumeCodeCacheTask. Using these new
functions, an embedder can coordinate the process of merging a newly
streamed script into a pre-existing script from the Isolate compilation
cache. The internal portions of this functionality have existed for a
while and are exercised when --stress-background-compile is set; this
change only plumbs the API calls. The new functionality is disabled by
default behind a runtime flag.

Bug: v8:12808
Change-Id: Ia7a14a739779ab658b505572d19df4ec489a078e