Fix update of evacuated objects in string forwarding table
Skip whole table entry when the original string is the deleted element
sentinel.
When a string dies in young space, only the original string object in
the forwarding table is overwritten with the deleted element sentinel to
mark the whole entry as deleted.
When updating objects after evacuation during full gc, rows weren't
correctly skipped.

Change-Id: I76bac70acdb174c0a48cf876b1527f9de048b2ef