[maglev] Fix OOB in StringFromCharCode
StringFromCharCode expects an int32 value, but maglev isn't careful
about keeping the top 32 bits of the register valid (to avoid needing to
sign extend after every 32-bit operation). This means the top bits of
an int32 register might be invalid when it is used -- in particular,
complex addressing uses its inputs as 64-bit values, including the
index.

Long story short, we need to zero the top bits of the int32 char_code
used as the index into the single character table.

Bug: v8:7700
Change-Id: I3540230c865a1d07c105f35d024d598cc8e15180
Fixed: chromium:1392585