[wasm] Speed up the JS WasmModuleBuilder
The WasmModuleBuilder is used in tests for creating Wasm modules. It can
be pretty slow for huge modules, in particular in simulator builds or in
slow variants like gc-stress.

This CL adds a fast path to the code section creation, for functions
without locals. This makes the wasm-max-functions test 1.45x faster in
the arm64 simulator (generation of the code section alone gets 2.2x
faster).

R=ahaas@chromium.org

Change-Id: I993542448fb4f0b5fdadca13c59691d86844e2a6