[maglev] Move verification code to maglev-ir.cc
Keep NodeT verification code closer to the reset of the NodeT
implementation by making input verification a function on NodeT rather
than being implemented in a separate verifier.

This function has a default implementation for fixed-input nodes, which
just looks up a static input types array (with CRTP) on the derive
node.

As a drive-by, change the fixed-size node base classes to share a mixin
base class, to avoid duplicating the verification logic.

Bug: v8:7700
Change-Id: Ib9dd7dba540f16fa359b170d513f20aa0e110e24