[heap] Fix DCHECK failure in NotifyFullSweepingCompleted
In case of nested GCs, NotifyFullSweepingCompleted will first notify
that the nested young cycle is finished, which in turn will check
whether we need stop the full cycle as well.
After returning from NotifyYoungSweepingCompleted,
NotifyFullSweepingCompleted tries to stop the full cycle again, which
triggers the DCHECK that the current cycle is in a sweeping state.

Bug: v8:12612
Change-Id: I659038c04533b71bbc379cd51d22441e44e96021