[heap] Disable black allocation in shared heap during client GC
A scavenger GC can run on a client isolate while incremental marking
is enabled in the shared space isolate. While we pause black
allocation in the client's heap, we can't really pause it in the
shared heap isolate.

This CL solves that by marking ConcurrentAllocator instances created
for the GC. This will probably also be useful when moving LABs out
of the spaces.

Bug: v8:13267
Change-Id: Ie92848854c39229a8eca74cb07bd4be19a725ec8