[wasm] Push growing of value stack into handlers
Previously (https://crrev.com/c/2431525), we moved checking the capacity
and growing the value stack into the central dispatch, to avoid doing
that repeatedly in the individual byte code handlers.

This change moves it into the handlers again, but also only checks and
grows the stack if it doesn't have already enough space from previous
operations (e.g., if a `Drop` happened just before a `Push`).

This change makes the dispatch code a bit smaller, which is good for a
tail call version of the decoder, where the dispatch code is inlined and
duplicated into each handler. In terms of runtime, it introduces about
0.25% overhead for single-threaded validation.

R=clemensb@chromium.org

Change-Id: Iaffd48e0c06f919ead17ad79ff740c42b10cff7d