[debug] Fix catch prediction for for-of statement
The bytecode generator currently uses the wrong catch prediction in
the desugaring for for-of loops. Specifically the call to .next of
the iterator is unconditionally marked as 'uncaught' instead of
using the surrounding catch prediction.

This leads to unexpected pauses in DevTools when 'pause on uncaught
exceptions' is enabled, e.g. for throwing generators.

R=leszeks@chromium.org

Bug: chromium:1270780
Change-Id: I2e642ef3fbfcfc6ad19e92cf611188801ebf2450