Remove GC post-processing check from externalizing strings
This check was probably relevant in the past (when GC was different),
but it is unknown today why it exists. It can cause races in a
concurrent set-up though, as tracking the post-processing state is not
atomic.
Due to possible races and no reason to check whether we are in GC
post-processing phase, this check is removed from
String::SupportsExternalization().

Drive-by: Remove tracking of GC post-processing state, as
SupportsExternalization() was the only user of the logic.
Change-Id: Id0a6dd25a8dc6044504b40f1c754612dedcacf75