[runtime] Zero padding of sequential strings
SeqStrings have up to kTaggedSize-1 bytes of padding, due to
allocation sizes being rounded up to kObjectAlignment. This CL ensures
that all (non-external) sequential strings on the heap have (and keep)
this padding set to all zeroes.

The approach is to unconditionally zero the last tagged word of the
uninitialized allocation of string objects.

Change-Id: I32ee4a53e25fa470f79562a71b8c648c7205523f