Fix stack corruption on register save on ppc64 systems
On the ppc64 platform, the register save assembler starts
saving registers at an offset of 1, not 0.  As a result,
the save space is overrun by one word, corrupting the
previous stack frame.  This causes either a crash or a
stack smashing alert depending on the build configuration.

Realign the saved registers to start with an offset of 0.

This fixes numerous Chromium crashes (pre-109) on ppc64,
along with a new FTBFS in the ppc64 version of Chromium 109.

Tested on a Talos II system under Debian Bullseye.

Change-Id: I276b6fb1689c7df59a822d27444672bfc868baf7