[static-roots] Check static map in instance type checks
A first stab at using some static maps to have faster instance type
checks in C++ code. This adds an instance type check variant which
uses the map without loading the instance_type field, when the
instance type corresponds to a single static map.

Changes the format of the static roots table to be more in line with
other code and orders the roots by address.

Bug: v8:13466
Change-Id: I5417071efd24ee52d51146ef0887d32cb9b62fcd