Decrease TranslationArray size by bit-packing opcode and first operand
This change implements an option for using only a single byte instead of
two for some common translation opcodes, if their operands can fit in
five bits. I see no difference in the time it takes to write translation
arrays, as measured by V8.TFCodeGeneration time, but the time to read
regresses slightly (about 1% on js-perf-test/StackTrace).

On Octane, this change reduces the total bytes of TranslationArrays
by 22% (270 kB).

Bug: v8:11354
Change-Id: I96917783b03a04523b024cfbaac0d222a39abb0a