[static-roots] Use static maps instead of instance type in CSA
If an instance type corresponds to a unique static map we do not need
to load the instance type from the map. Instead we can compare the
compressed map word against the known map.

This CL addresses only the backend. We'll add the same optimization to
the CodeStubAssembler in a follow up.

Bug: v8:13466
Change-Id: I3a53299e0475e90eb289a23f09e7374dfc10bad1